requirements.txt
# Core dependencies
opencv-contrib-python>=4.9.0
numpy>=1.26.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.6.0
colorama>=0.4.6

# Object detection (CoreML - no PyTorch required at runtime)
coremltools>=7.0

# Optional: Export YOLO to CoreML (one-time, run scripts/export_coreml.py)
# ultralytics>=8.0.0
# torch>=2.0.0
